feat(src/common/components/addons/LazyLoad): implement LazyLoad addon component
